An attempt to improve the performance of routing.
This will be reverted, but I'm adding it here for reference.

The overhead comes from creating the map. I still want to find a way to reduce the problem search space after each recursion, but it will have to be a very cheap way.

Before:
BenchmarkStarRoute           500           5527091 ns/op

After:
BenchmarkStarRoute      1000           2257257 ns/op